requirements.txt
pyodbc
SQLAlchemy
azure-identity
fastapi
uvicorn
mcp
fastapi-mcp
pandas
python-dotenv
requests
tabulate